Revert "core: correctly parse boolean configuration values."
This reverts commit 8dd1c2d4aff5e40e4a87f59fa09f8de3f2232fc1.

Introduced a security bug: since CCAN/opt is used to parse the JSON
config (something it is not meant to do), it was possible to specify
the default of

    "api-listen":false

in .conf and expect the API to be disabled. However, since CCAN/opt
does not check for parameters to OPT_WITHOUT_ARG, this got set to
true anyway, and enabled the API on a (possibly unfirewalled) network.

For this reason, configuration options are not named uniformly.

This should have been reverted as soon as it became known:

4c4b909be6dd8643372ec683ec73f73725cb3a77 (Fri Feb 28)

Sorry for that.